Delete campaign tag and replace it by generic post tag
